+++ src/components/CustomCodeEditor/CustomCodeEditor.test.tsx
@@ -34,7 +34,7 @@ describe('Custom Code Editor', () => {
    * @param item
    * @param restProps
    */
-  const getComponent = ({ value = [], item = {}, ...restProps }: any) => {
+  const getComponent = ({ value = '', item = {}, ...restProps }: any) => {
     return <CustomCodeEditor {...restProps} value={value} item={item} />;
   };
 
@@ -87,6 +87,7 @@ describe('Custom Code Editor', () => {
         monacoOptions: {
           formatOnPaste: true,
           formatOnType: true,
+          scrollBeyondLastLine: false,
         },
       }),
       expect.anything()
@@ -166,7 +167,7 @@ describe('Custom Code Editor', () => {
       () =>
         ({
           getVariables: jest.fn().mockImplementation(() => variables),
-        } as any)
+        }) as any
     );
 
     /**
@@ -220,7 +221,7 @@ describe('Custom Code Editor', () => {
       () =>
         ({
           getVariables: jest.fn().mockImplementation(() => variables),
-        } as any)
+        }) as any
     );
 
     /**